Ui and crud details completed

+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Validator;
 
@@ -84,17 +85,19 @@ class UserController extends Controller
 
         $user = User::find($userId);
 
+        if ($user->id == Auth::user()->id) {
+            return 'Impossible to edit the logged user';
+        }
+
         $data = $request->validate([
             'name' => ['required', 'string', 'max:255'],
             'email' => ['required', 'string', 'email', 'max:255', 'unique:users,email,' . $user->id],
             'password' => ['string', 'min:8'],
         ]);
 
-        $user->update([
-            'name' => $data['name'],
-            'email' => $data['email'],
-            'password' => Hash::make($data['password'])
-        ]);
+        $data['password'] = empty($request->password) ? $user->password : Hash::make($request->password);
+
+        $user->update($data);
     }
 
     /**
@@ -106,6 +109,10 @@ class UserController extends Controller
     public function destroy($userId)
     {
         $user = User::find($userId);
+
+        if ($user->id == Auth::user()->id) {
+            return 'Impossible to delete the logged user';
+        }
 
         $user->delete();
     }
